# Dashboard: Update stories per request count to speed up time to first interaction