## Summary
I was a bit hasty in my update to set max stories per the original conversation attached to this PR's story :) Turns out that really slows things down! Per conversation with Omar, 24 stories per request is a happy medium for users with huge monitors and then the average user on a laptop or a tablet who is going to see max 10 stories (2 row of 5 at 1400px which is still large). 

## Relevant Technical Choices

**100 stories**
Stories API
50ish Mbps @ 100 stories per request API time takes 434 - 497ms to load 
3G slow @ 100 stories per request API time takes about 42s to load

Total timing
50ish Mbps:  Finish 5.9s, DOMContentLoaded 3.21s, Load 3.62s 
3G slow: Finish 1.8 min, DOMContentLoaded 44.30s, Load 47.43s

**60 Stories**  
Stories API 
50ish Mbps @ 60 stories per request API time takes 344 - 376ms to load 
3G slow @ 60 stories per request API time takes about 28s to load

Total timing
50ish Mbps:  Finish 1.54s, DOMContentLoaded 1.19s, Load 1.22s 
3G slow: Finish 1.2 min, DOMContentLoaded 44.39s, Load 47.43s

**24 Stories**  
Stories API 
50ish Mbps @ 24 stories per request API time takes about 203ms to load 
3G slow @ 24 stories per request API time takes about 13.11s to load

Total timing 
50ish Mbps:  Finish 1.38s, DOMContentLoaded 1.19s, Load 1.22s 
3G slow: Finish 57.34s, DOMContentLoaded 44.23s, Load 47.77s

## To-do
Deeper dive into perf 

## User-facing changes
- Dashboard should load faster

## Testing Instructions

- Verify that Dashboard loads faster 

---

<!-- Please reference the issue(s) this PR addresses. -->

Addresses #2076