@BrittanyIRL 2020-06-29
Hey @swissspidy - do you mind taking a quick look and approving this if you're good with it. I'm having trouble getting the 'merge' button to activate. Omar QA'd this for me, see thread here: https://github.com/google/web-stories-wp/issues/2076